requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
azure-core==1.30.1
azure-storage-blob==12.19.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.14.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
isodate==0.6.1
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
